IAT-440: Refactor core services into core module.
