+++ src/app/core/alert.service.spec.ts
@@ -1,6 +1,4 @@
 import {inject, TestBed} from "@angular/core/testing";
-import {HttpModule} from "@angular/http";
-import {Logger} from "./logger.service";
 import {AlertService} from "./alert.service";
 import {ToastyModule} from "ng2-toasty";
 
@@ -8,12 +6,10 @@ describe('AlertService', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [
-        HttpModule,
         ToastyModule
       ],
       providers: [
-        AlertService,
-        Logger
+        AlertService
       ]
     });
   });
