+++ src/app/core/logger.service.spec.ts
@@ -1,0 +1,17 @@
+import {inject, TestBed} from "@angular/core/testing";
+import {Logger} from "./logger.service";
+
+describe('LoggerService', () => {
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      imports: [],
+      providers: [
+        Logger
+      ]
+    });
+  });
+
+  it('should ...', inject([Logger], (service: Logger) => {
+    expect(service).toBeTruthy();
+  }));
+});
