+++ src/app/item/item-create/item-create.component.ts
@@ -17,7 +17,7 @@ import "rxjs/add/operator/switchMap";
 import {Component, OnInit} from "@angular/core";
 import {ActivatedRoute, Router} from "@angular/router";
 import {ItemService} from "../../service/item.service";
-import {Logger} from "../../service/logger.service";
+import {Logger} from "../../core/logger.service";
 import {Item} from "../../model/item/item";
 
 @Component({
