+++ src/app/item/item-create/item-select-type/item-select-type.component.spec.ts
@@ -3,7 +3,7 @@ import { HttpModule } from '@angular/http';
 import { ItemSelectTypeComponent } from './item-select-type.component';
 import { LookupService} from '../../../service/lookup.service';
 import { RouterTestingModule} from '@angular/router/testing';
-import {Logger} from "../../../service/logger.service";
+import {Logger} from "../../../core/logger.service";
 
 describe('ItemSelectTypeComponent', () => {
