+++ src/app/item/item-create/item-select-type/item-select-type.component.ts
@@ -17,7 +17,7 @@ import {Component, OnInit} from "@angular/core";
 import {Router} from "@angular/router";
 import {LookupService} from "../../../service/lookup.service";
 import {ItemType} from "../../../model/item-type";
-import {Logger} from "../../../service/logger.service";
+import {Logger} from "../../../core/logger.service";
 
 @Component({
   selector: 'app-item-select-type',
