+++ src/app/item/item-load-sa/item-load-sa.component.ts
@@ -15,7 +15,7 @@
  */
 import {AfterViewChecked, Component, ElementRef, Input, OnInit} from "@angular/core";
 import {FormArray, FormBuilder, FormGroup} from "@angular/forms";
-import {Logger} from "../../service/logger.service";
+import {Logger} from "../../core/logger.service";
 import {Item} from "../../model/item/item";
 import {SaItem} from "../../model/item/sa-item";
