+++ src/app/item/item-load/item-load.component.spec.ts
@@ -6,8 +6,8 @@ import { AlertModule } from 'ngx-bootstrap/alert';
 import { ModalModule } from 'ngx-bootstrap/modal';
 import { ItemLoadSaComponent} from '../item-load-sa/item-load-sa.component';
 import { ItemLoadComponent } from './item-load.component';
-import {Logger} from "../../service/logger.service";
-import {AlertService} from "../../service/alert.service";
+import {Logger} from "../../core/logger.service";
+import {AlertService} from "../../core/alert.service";
 import {ToastyModule} from "ng2-toasty";
 
 describe('ItemLoadComponent', () => {
